1.添加es查询 2.移除原content表 3.统一分页前端参数 4.支持文章模糊查询 5.修改常量 6修改swagger信息

+++ src/main/java/cn/lq/common/domain/dto/StatisticsDto.java
@@ -13,7 +13,7 @@ public class StatisticsDto implements Serializable {
     /**
      * 文章数
      */
-    private Integer articles;
+    private Long articles;
     /**
      * 评论数
      */
@@ -27,11 +27,11 @@ public class StatisticsDto implements Serializable {
      */
     private Integer attachs;
 
-    public Integer getArticles() {
+    public Long getArticles() {
         return articles;
     }
 
-    public void setArticles(Integer articles) {
+    public void setArticles(Long articles) {
         this.articles = articles;
     }
